render: decrease RrRect to 12 bytes (1 cpu cacheline)

+++ obrender/geom.h
@@ -26,8 +26,8 @@ typedef struct {
 } RrSize;
 
 typedef struct {
-    int x;
-    int y;
+    short x;
+    short y;
     int width;
     int height;
 } RrRect;
